Move `Back` button next to `Play Game` button
This commit moves the `Back` and `Play Game` buttons next to one another instead of below, I believe this improves the look and flow of the web page.

Also, the `New Game` button in the checkmate modal was fixed by assigning it a different id than the already existing `New Game` button under the chessboard.