Fix `hard` mode not responding
There was another issue with `hard` mode: While we set a depth limit in the `/get_evaluation` route, there was no limit set in the `/get_ai_move` route, which again prevented hard mode from making any moves.

This change was probably reversed in a merge that did not get properly resolved.